requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
shap>=0.42
chromadb>=0.4
sentence-transformers>=2.2
fastapi>=0.100
uvicorn>=0.23
pyyaml>=6.0
plotly>=5.15
pytest>=7.4
httpx>=0.24
